test(cli): extract priorities play tests
Move the remaining GamePlayPriorities coverage into packages/cli/test/commands/game/play/priorities.test.ts and remove the emptied game.play.test.ts monolith from test:cli:play.

Record the parallel priorities agent handoff in the support modularization OpenSpec workstream: the agent produced a boundary-clean net-new candidate, while DRA retained single-writer ownership for package.json and monolith deletion.

Validation: bun run --cwd packages/cli test -- game/play/priorities.test.ts; bun run --cwd packages/cli test -- game/play/priorities.test.ts game/play/notification/hud.test.ts game/play/ready-city.test.ts game/play/ready-unit.test.ts game.play.tactical-read.test.ts -t 'priorit|surfaces|routes|classifies|reads ready|relationship'; bun run check:cli; bun run test:cli:play; bun run openspec -- validate civ7-support-direct-control-modularization --strict; git diff --check.
