+++ openspec/changes/civ7-support-direct-control-modularization/workstream/review-disposition-ledger.md
@@ -11,3 +11,4 @@
 | oRPC authority/source lane is unresolved in current support branch | `019e8ac3-b2ca-7b53-910a-bc6286c4d04b`, `019e8ac9-296c-74a3-9d30-f4e3a4f51545` | P2 | Accepted/deferred | `.agents/skills/civ7-orpc-control-architecture/SKILL.md` is absent here; `codex/civ7-orpc-control-architecture-skill` and `add-control-orpc-*` branches exist. Effect/oRPC stays planning-only until source/skill authority is imported or cited. |
 | Future delegations require durable close-loop discipline | user correction and supervisor `019e8a79-aeee-7dd3-9fd1-1ef616078ef4` | P1 | Accepted/repaired | Workstream record now contains thread ledger with ids, wave, read status, status, and disposition. |
 | HUD parallel implementation lane must not be treated as proof-complete from the agent worktree | `019e8ad7-dc15-76a0-88f8-bb5210bfd7e9` and supervisor `019e8a79-aeee-7dd3-9fd1-1ef616078ef4` | P1 | Accepted/repaired | HUD agent output was recorded as a boundary-clean candidate only; final proof is owned by the DRA worktree focused/adjacent/check/play/OpenSpec gates. |
+| Priorities parallel implementation lane must not be treated as proof-complete from the agent worktree | `019e8ae4-8b20-79a0-8553-41b71bccb63f` and supervisor `019e8a79-aeee-7dd3-9fd1-1ef616078ef4` | P1 | Accepted/repaired | Priorities agent output was recorded as a boundary-clean candidate only; final proof is owned by the DRA worktree focused/adjacent/check/play/OpenSpec gates. |
